// app-angular-nest/client/src/app/core/menu/menu.model.ts
export interface IMenuItem {
  id: number | string;
  type: 'section' | 'link' | 'sub' | 'extLink' | 'extTabLink';
  name: string;
  icon?: string;
  route: string;
  child?: IMenuChildrenItem[];
  permissions?: IMenuPermissions;
}

export interface IMenuChildrenItem extends IMenuItem{
  type: 'link' | 'sub' | 'extLink' | 'extTabLink';
}

export interface IMenuPermissions {
  only?: string | string[];
  except?: string | string[];
}